Added timestamp formatting and moved classes back into their respective java files
