+++ src/ui/SearchView.java
@@ -17,7 +17,6 @@ import java.awt.Insets;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.awt.image.BufferedImage;
-import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -152,6 +151,75 @@ public class SearchView {
      */
     public JComponent $$$getRootComponent$$$() {
         return basePane;
+    }
+
+    /**
+     * Created by Clay on 4/12/2016.
+     */
+    static class ResultListObject {
+        private ImageIcon icon;
+        private User user;
+
+        public ResultListObject(User u) {
+            user = u;
+            BufferedImage myPicture = null;
+            try {
+                ArrayList<String> images = UserPhotosTable.getUserPhotos(RelationshipController.getConnection(), user);
+                if (images.size() == 0) {
+                    myPicture = ImageIO.read(new File("resources/images/logo.png"));
+                } else {
+
+                    URL url = new URL(images.get(0));
+                    myPicture = ImageIO.read(url);
+                }
+
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            //resize
+            double factor = (double) 100 / (double) myPicture.getHeight();
+
+            Image newimg = myPicture.getScaledInstance((int) (myPicture.getWidth() * factor), (int) (myPicture.getHeight() * factor), Image.SCALE_SMOOTH);
+            //
+            icon = new ImageIcon(newimg);
+        }
+
+        public ImageIcon getIcon() {
+            return icon;
+        }
+
+        public String getName() {
+            return user.getUsername();
+        }
+    }
+
+    /**
+     * Created by Clay on 4/12/2016.
+     */
+    static class UserListRenderer extends JLabel implements ListCellRenderer {
+        private static final Color HIGHLIGHT_COLOR = new Color(88, 130, 255);
+
+        public UserListRenderer(RelationshipController controller) {
+            setOpaque(true);
+            setIconTextGap(12);
+        }
+
+        public Component getListCellRendererComponent(JList list, Object value,
+                                                      int index, boolean isSelected, boolean cellHasFocus) {
+            ResultListObject entry = (ResultListObject) value;
+            setText(entry.getName());
+
+            setIcon(entry.getIcon());
+            if (isSelected) {
+                setBackground(HIGHLIGHT_COLOR);
+                setForeground(Color.white);
+            } else {
+                setBackground(Color.white);
+                setForeground(Color.black);
+            }
+            return this;
+        }
     }
 }
